Size polices and general component sizing recode     moved shapes to utilities     brought back size polices, yeeeeeeeaa     added size policy interface in utilities     created new size policies package with some common size policy implementations     renamed extendsd content button to custom button     image and text content button improvements     added width and height to image - user set value     fixed drawable content wrong constructor parameter type (from int to double)     polished some components sizing     fixed combo box click on text to open choices     fixed slider click to set value

+++ cz/mg/toolkit/component/DrawableContent.java
@@ -13,7 +13,7 @@ public abstract class DrawableContent extends Content implements DrawableCompone
         addEventListeners();
     }
     
-    public DrawableContent(int contentWidth, int contentHeight) {
+    public DrawableContent(double contentWidth, double contentHeight) {
         super(contentWidth, contentHeight);
         addEventListeners();
     }
